Changed collect to return a roaring bitmap instead of a vector

+++ src/database.rs
@@ -381,7 +381,9 @@ impl Database {
                 total_collect_indices_time += rle_collect_indices_start.elapsed().as_secs_f64();
 
                 let increment_counts_start = Instant::now();
-                indicies.into_iter().for_each(|index| num_hits[index] += 1);
+                indicies
+                    .into_iter()
+                    .for_each(|index| num_hits[index as usize] += 1);
                 total_increment_counts_time += increment_counts_start.elapsed().as_secs_f64();
             }
             // Increment the total number of queries
